Extract NominalMap fields to internal structs
This change also temporarily removes the name_to_instances field
in the NominalMap. As for why, I haven't really found a use for
it yet. I think it makes more sense to implement it downstream
where it's able to use a fully resolved class name key, which can
enable "implementation" lookup for a class.