Allocate IDs on-the-fly for instance variables
This implements the idea of allocating IDs on-the-fly to serve as
anchors for instance variables that are implicitly quantified. Instead
of a simple HashSet, GraphNode::Constraint uses an Interner to provide
stable IDs for each type variable to be used for the De Bruijn context.

However, these IDs are not tied to the source map, so it's non-trivial
to determine their source positions unlike TypeVariableBindingId.