Make ResolvedImports more fit for resolution
Previously, ResolvedImports only contained the names of whatever was
specified in the import list. This made it impossible to "lookup"
'implicit' imports brought about by 'hidden' imports.

By changing the definition of ResolvedImports to contain all items,
tagged with an ImportKind, it's now possible to perform lookups against
all kinds of imports for the purposes of name resolution.