Add checking rules in Arrow and Forall
For Arrow, this previously meant that any `argument` or `result`
that had a unification variable for its kind wasn't properly
checked against the `Type` kind, same goes for the `inner` in
Forall.

This effect can be observed prominently in the foreign_poly test,
where the signatures for TuplePoly and InferTuplyPoly are simplified.